created universal header/footer, created new file for national data processing functions, rendered all app templates
