+++ templates/footer.html
@@ -1,0 +1,11 @@
+<footer class="page-footer">
+      <div class="row" style="background:linear-gradient(90deg,#00ade6 0,#05e3d5)">
+          <div class="col-12" style="height:5px">
+          </div>
+      </div>
+        <div class="row" style="background-color:black; height:140px">
+            </div>
+        </div>
+    </div>
+  </div>
+</footer>
