+++ templates/header.html
@@ -1,0 +1,24 @@
+<nav class="navbar navbar-toggleable-sm navbar-light bg-faded navbar-full">
+    <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+    </button>
+    <a class="nav-link" id=homeButton href="index.html">Functional_Vice()</a>
+    <div class="collapse navbar-collapse" id="navbarNavDropdown">
+        <ul class="navbar-nav ml-auto">
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" href="" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                    Analysis
+                </a>
+                <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
+                    <a class="dropdown-item" href="/">Details</a>
+                </div>
+            </li>
+            <li class="nav-item">
+                    <a class="nav-link" href="/gauges_tweets">Correlation</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="">Data</a>
+            </li>
+        </ul>
+    </div>
+</nav>
